moved all component files to components folder
